设计实现2：核心流程设计分析/my_memo/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
beautifulsoup4=4.12.2=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.08.22=haa95532_0
libffi=3.4.4=hd77b12b_0
openssl=3.0.12=h2bbff1b_0
pillow=10.1.0=pypi_0
pip=23.3=py311haa95532_0
pyqt5=5.15.10=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.13.0=pypi_0
python=3.11.5=he1021f5_0
setuptools=68.0.0=py311haa95532_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h2bbff1b_0
tk=8.6.12=h2bbff1b_0
tzdata=2023c=h04d1e81_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.41.2=py311haa95532_0
xz=5.4.2=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
